Split objects into seperate folders and created a bibtex file for citations
Since several nodes and edges may cite the same source it would be wise
to put them all into a seperate bibtex file and just reference the name
of the citation.  Additionally the "url" field is deprecated since the
url should be with the citation in the bibtex file under "howpublished".
  source_node is replaced with source_nodes which is an array of IDs so
an edge can reference multiple nodes as inputs.